+++ core/plugin/3d_layer/index.d.ts
@@ -1,3 +1,0 @@
-import { IPluginInitArg } from '@famibee/skynovel';
-export declare function init(pia: IPluginInitArg): Promise<void>;
-//# sourceMappingURL=index.d.ts.map