+++ core/plugin/cubism3_layer/index.js
@@ -1,9 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.init = init;
-const Cubism3Layer_1 = require("./Cubism3Layer");
-async function init(pia) {
-    pia.addLayCls('cubism3', () => new Cubism3Layer_1.Cubism3Layer);
-}
-;
-//# sourceMappingURL=index.js.map