+++ core/web4webpack_s.js
@@ -1,6 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-// 変更後は「npm run webpack:dev」
-
-import {SysWeb} from '@famibee/skynovel/web';
-new SysWeb({}, {cur: 'prj/'});